Initial implementation of mode switching
Mode switching is implemented as a JDEECo plugin (ModeSwitchingPlugin). 
Each DEECo component should have an annotation specifying the modechart
class that governs its mode-switching. Each process inside the component
that is subject to mode-switching should have an annotation specifying
the mode it belongs to.
Mode-switching is performed via a dedicated process of a "System" DEECo
component, the ModeSwitchingManager, which performs mode-switching for
all the components deployed on its deeco node.